Separate notify server status & allele validation, so we can disable allele validation in non-Shariant environments
